Fix usage of `re.match` in `detector.py` - most should have been `re.search`.
`re.match` silently adds the requirement that the regex match from the start of the given string, whereas `re.search` does not enforce this (although specifying `^` in the regex will still match only the start of a string). Given this, several uses of `re.match` should be `re.search`, including the code that checks for robotic user agents (which wasn't working).